Extract `updatable-cache-item` files
