+++ tsconfig.json
@@ -6,9 +6,11 @@
 		"checkJs": true
 	},
 	"files": [
+		"updatable-cache-item.test-d.ts",
+		"updatable-cache-item.ts",
 		"cache-item.test-d.ts",
 		"cache-item.ts",
 		"index.test-d.ts",
-		"index.ts"
+		"index.ts",
 	]
 }
